Error fixes + add vehicle functionality
